requirements.txt
aiosqlite
discord.py
gTTS
loguru
uvloop
PyNaCl
jishaku
